# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SkillForge AI - Learning Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules/{moduleId}:
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Update module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Lessons
  /modules/{moduleId}/lessons:
    get:
      operationId: getLessons
      tags: [lessons]
      summary: List lessons for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create lesson
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Enrollments
  /enrollments:
    get:
      operationId: getEnrollments
      tags: [enrollments]
      summary: Get enrollments for current user
      responses:
        "200":
          description: Enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: enrollCourse
      tags: [enrollments]
      summary: Enroll in a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollInput"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{enrollmentId}:
    delete:
      operationId: unenrollCourse
      tags: [enrollments]
      summary: Unenroll from course
      parameters:
        - name: enrollmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unenrolled
...
  /messages/conversations:
    get:
      operationId: getConversations
      tags: [messages]
      summary: Get conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [messages]
      summary: Start a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /messages/conversations/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # Course discussions
  /courses/{courseId}/discussions:
    get:
      operationId: getDiscussions
      tags: [messages]
      summary: Get course discussions
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discussions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discussion"
    post:
      operationId: createDiscussion
      tags: [messages]
      summary: Post to discussion
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"

  # Dashboard endpoints
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, instructor, admin]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Course:
      type: object
      required: [id, title, slug, instructorId, categoryId, level, isPublished, price, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        previewVideoUrl:
          type: ["string", "null"]
        instructorId:
          type: integer
        instructorName:
          type: ["string", "null"]
        instructorAvatar:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        isPublished:
          type: boolean
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        enrollmentCount:
          type: integer
        totalLessons:
          type: integer
        totalDuration:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Module:
      type: object
      required: [id, courseId, title, position]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        position:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
...
    ModuleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        position:
          type: integer

    # Lessons
    Lesson:
      type: object
      required: [id, moduleId, title, type, position]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [video, text, quiz, assignment, resource]
        content:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        resourceUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        position:
          type: integer
        isFree:
          type: boolean
        isCompleted:
          type: boolean
    LessonInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [video, text, quiz, assignment, resource]
        content:
          type: string
        videoUrl:
          type: string
        resourceUrl:
          type: string
        duration:
          type: integer
        position:
          type: integer
        isFree:
          type: boolean
    LessonUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [video, text, quiz, assignment, resource]
        content:
          type: string
        videoUrl:
          type: string
        resourceUrl:
          type: string
        duration:
          type: integer
        position:
          type: integer
        isFree:
          type: boolean

    # Enrollments
    Enrollment:
      type: object
      required: [id, userId, courseId, enrolledAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        course:
          $ref: "#/components/schemas/Course"
        enrolledAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
        progressPercent:
          type: number
    EnrollInput:
      type: object
      required: [courseId]
      properties:
        courseId:
          type: integer

    # Progress
...
    Conversation:
      type: object
      required: [id, participants, lastMessage, createdAt]
      properties:
        id:
          type: integer
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ConversationInput:
      type: object
      required: [participantId]
      properties:
        participantId:
          type: integer
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderAvatar:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    Discussion:
      type: object
      required: [id, courseId, userId, content, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        userRole:
          type: string
        content:
          type: string
        parentId:
          type: ["integer", "null"]
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Discussion"
        createdAt:
          type: string
          format: date-time
    DiscussionInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentId:
          type: integer

    # Dashboard
...